Throw NoSuchMethodException when Constructor is missing
In the original code `Clazz.searchAndExecuteMethod` does not throw an
exception when a constructor is not found, but silently does nothing.

Throwing an exception in case a constructor is missing is important as
in our code reflection is used to tryout various constructors. With the
original behaviour that reflection code was not working because it
assumed the constructor was found and ran into an error.

A similar change is applied when calling super does not find the
constructor (see `Clazz.superCall` in `j2slib.z.js`).

#### Fix (in `j2slib.z.js`)

*(Beautified code)*

    Clazz.superCall = function(To, Tz, Nf, funParams) {
        ...
        if (fx != null) {
            ...
        } else if (Nf == "construct") {
            throw new NoSuchMethodException();
        }
        throw new Clazz.MethodNotFoundException(To, Tz, Nf, Clazz.getParamsType(funParams).typeString);
    };