Log a warning to console when SAEM cannot find a function
When Clazz.searchAndExecuteMethod cannot find a function it quietly
returned, leaving no clue what went wrong.

In the related comment it say not finding a function may be caused by a
missing default constructor. In that case doing nothing would be fine.

However we never encountered the "missing default constructor" case but
problems caused by bugs in the J2S runtime.

Logging a warning when a function is not found gives the developer a
hint where to look (when in debugging mode). In case of the mentioned
"default constructor" things just work as before (for the end user) as
the log information is only visible when debugging.